+++ app/org/[org_id]/bridges/configure/[id]/page.js
@@ -5,24 +5,28 @@ import Chat from "@/components/configuration/chat";
 import Chatbot from "@/components/configuration/chatbot";
 import LoadingSpinner from "@/components/loadingSpinner";
 import Protected from "@/components/protected";
-import { useCustomSelector } from "@/customSelector/customSelector";
-import { createApiAction, getSingleBridgesAction, integrationAction } from "@/store/action/bridgeAction";
+import { useCustomSelector } from "@/customHooks/customSelector";
+import { getSingleBridgesAction } from "@/store/action/bridgeAction";
 import { getModelAction } from "@/store/action/modelAction";
-import { useEffect } from "react";
+import { useEffect, useRef } from "react";
 import { useDispatch } from "react-redux";
 
 export const runtime = 'edge';
 const Page = ({ params }) => {
+  const mountRef = useRef(false);
   const dispatch = useDispatch();
-  const { bridgeType, service, embedToken, isServiceModelsAvailable } = useCustomSelector((state) => ({
+  const { bridgeType, service, isServiceModelsAvailable } = useCustomSelector((state) => ({
     bridgeType: state?.bridgeReducer?.allBridgesMap?.[params?.id]?.bridgeType,
     service: state?.bridgeReducer?.allBridgesMap?.[params?.id]?.service,
-    embedToken: state?.bridgeReducer?.org?.[params?.org_id]?.embed_token,
     isServiceModelsAvailable: state?.modelReducer?.serviceModels?.[state?.bridgeReducer?.allBridgesMap?.[params?.id]?.service],
   }));
 
   useEffect(() => {
     dispatch(getSingleBridgesAction(params.id));
+    return () => {
+      if (handleclose)
+        handleclose();
+    }
   }, []);
 
   useEffect(() => {
@@ -32,52 +36,23 @@ const Page = ({ params }) => {
   }, [service]);
 
   useEffect(() => {
-    if (embedToken) {
-      const script = document.createElement("script");
-      script.setAttribute("embedToken", embedToken);
-      script.id = process.env.NEXT_PUBLIC_EMBED_SCRIPT_ID;
-      script.src = process.env.NEXT_PUBLIC_EMBED_SCRIPT_SRC;
-      document.body.appendChild(script);
-
-      return () => {
-        document.body.removeChild(document.getElementById(process.env.NEXT_PUBLIC_EMBED_SCRIPT_ID));
-      };
-    }
-  }, [embedToken]);
-
-  useEffect(() => {
-    window.addEventListener("message", handleMessage);
-
-    return () => {
-      window.removeEventListener("message", handleMessage);
-    };
-  }, [params.id]);
-
-  useEffect(() => {
     handleResizer();
   }, []);
 
-  function handleMessage(e) {
-    // todo: need to make api call to update the name & description
-    if (e?.data?.webhookurl) {
-      const dataToSend = {
-        ...e.data,
-        status: e?.data?.action
-      }
-      dispatch(integrationAction(dataToSend, params?.org_id));
-      if ((e?.data?.action === "published" || e?.data?.action === "paused" || e?.data?.action === "created") && e?.data?.description?.length > 0) {
-        const dataFromEmbed = {
-          url: e?.data?.webhookurl,
-          payload: e?.data?.payload,
-          desc: e?.data?.description,
-          id: e?.data?.id,
-          status: e?.data?.action,
-          title: e?.data?.title,
-        };
-        dispatch(createApiAction(params.org_id, dataFromEmbed));
+  useEffect(() => {
+    if (mountRef.current) {
+      if (bridgeType === 'chatbot') {
+        if (typeof openChatbot !== 'undefined') {
+          openChatbot()
+        }
+      } else {
+        if (typeof closeChatbot !== 'undefined') {
+          closeChatbot()
+        }
       }
     }
-  }
+    mountRef.current = true;
+  }, [bridgeType])
 
   return (
     <>
@@ -92,7 +67,9 @@ const Page = ({ params }) => {
             <div className="resizer w-full md:w-1 bg-base-500 cursor-col-resize hover:bg-primary"></div>
             <div className="w-full md:w-1/3 flex-1 chatPage min-w-[450px]">
               <div className="p-4 m-10 md:m-0 h-auto lg:h-full" id="parentChatbot" style={{ minHeight: "85vh" }}>
-                {bridgeType === 'chatbot' ? <Chatbot params={params} /> : <Chat params={params} />}
+                {/* {bridgeType === 'chatbot' ? <Chatbot params={params} /> : <Chat params={params} />} */}
+                <Chatbot params={params} />
+                <Chat params={params} />
               </div>
             </div>
           </div>
