+++ components/chatbot/chatbotCard.js
@@ -1,4 +1,4 @@
-import { useCustomSelector } from "@/customSelector/customSelector"
+import { useCustomSelector } from "@/customHooks/customSelector"
 import { ArrowUpRight } from "lucide-react"
 import { useRouter } from "next/navigation"
 import { useState } from "react";
