+++ components/configuration/ConfigurationPage.js
@@ -1,4 +1,4 @@
-import { useCustomSelector } from "@/customSelector/customSelector";
+import { useCustomSelector } from "@/customHooks/customSelector";
 import { Bot, Cog } from "lucide-react";
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useState } from 'react';
