+++ components/configuration/chatTextInput.js
@@ -1,5 +1,5 @@
 import { dryRun } from '@/config';
-import { useCustomSelector } from '@/customSelector/customSelector';
+import { useCustomSelector } from '@/customHooks/customSelector';
 import _ from 'lodash';
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
