+++ components/configuration/configurationComponent/bridgeNameInput.js
@@ -1,4 +1,4 @@
-import { useCustomSelector } from '@/customSelector/customSelector';
+import { useCustomSelector } from '@/customHooks/customSelector';
 import { updateBridgeAction } from '@/store/action/bridgeAction';
 import { Pencil } from 'lucide-react';
 import React, { useCallback } from 'react';
