+++ components/configuration/configurationComponent/bridgeTypeToggle.js
@@ -1,4 +1,4 @@
-import { useCustomSelector } from '@/customSelector/customSelector';
+import { useCustomSelector } from '@/customHooks/customSelector';
 import { updateBridgeAction } from '@/store/action/bridgeAction';
 import { Info } from 'lucide-react';
 import React from 'react';
