+++ components/configuration/configurationComponent/embedListSuggestionDropdownMenu.js
@@ -1,4 +1,4 @@
-import { useCustomSelector } from '@/customSelector/customSelector';
+import { useCustomSelector } from '@/customHooks/customSelector';
 import { getStatusClass } from '@/utils/utility';
 import { Plus } from 'lucide-react';
 import React, { useMemo } from 'react';
