+++ components/configuration/configurationComponent/functionParameterModal.js
@@ -1,4 +1,4 @@
-import { useCustomSelector } from '@/customSelector/customSelector';
+import { useCustomSelector } from '@/customHooks/customSelector';
 import { parameterTypes } from '@/jsonFiles/bridgeParameter';
 import { updateBridgeAction } from '@/store/action/bridgeAction';
 import { Info, Trash2 } from 'lucide-react';
