+++ components/configuration/configurationComponent/responseFormatSelector.js
@@ -1,6 +1,6 @@
 "use client";
 
-import { useCustomSelector } from '@/customSelector/customSelector';
+import { useCustomSelector } from '@/customHooks/customSelector';
 import { updateBridgeAction } from '@/store/action/bridgeAction';
 import { isValidJson, validateUrl } from '@/utils/utility';
 import { useEffect, useState } from 'react';
