+++ components/configuration/configurationComponent/richTextToggle.js
@@ -1,4 +1,4 @@
-import { useCustomSelector } from '@/customSelector/customSelector';
+import { useCustomSelector } from '@/customHooks/customSelector';
 import { updateBridgeAction } from '@/store/action/bridgeAction';
 import React from 'react'
 import { useDispatch } from 'react-redux';
