+++ components/historyPageComponents/assistFile.js
@@ -1,21 +1,5 @@
 import { useEffect } from "react";
 
-export const useEmbedScriptLoader = (embedToken) => {
-  useEffect(() => {
-    if (embedToken) {
-      const script = document.createElement("script");
-      script.setAttribute("embedToken", embedToken);
-      script.id = process.env.NEXT_PUBLIC_EMBED_SCRIPT_ID;
-      script.src = process.env.NEXT_PUBLIC_EMBED_SCRIPT_SRC;
-      document.body.appendChild(script);
-
-      return () => {
-        document.body.removeChild(document.getElementById(process.env.NEXT_PUBLIC_EMBED_SCRIPT_ID));
-      };
-    }
-  }, [embedToken]);
-};
-
 export const useCloseSliderOnEsc = (setIsSliderOpen) => {
   useEffect(() => {
     const closeSliderOnEsc = (event) => {
