+++ components/historyPageComponents/embedScriptLoader.js
@@ -1,21 +1,21 @@
-import { useEffect } from "react";
+// import { useEffect } from "react";
 
-const EmbedScriptLoader = ({ embedToken }) => {
-  useEffect(() => {
-    if (embedToken) {
-      const script = document.createElement("script");
-      script.setAttribute("embedToken", embedToken);
-      script.id = process.env.NEXT_PUBLIC_EMBED_SCRIPT_ID;
-      script.src = process.env.NEXT_PUBLIC_EMBED_SCRIPT_SRC;
-      document.body.appendChild(script);
+// const EmbedScriptLoader = ({ embedToken }) => {
+//   useEffect(() => {
+//     if (embedToken) {
+//       const script = document.createElement("script");
+//       script.setAttribute("embedToken", embedToken);
+//       script.id = process.env.NEXT_PUBLIC_EMBED_SCRIPT_ID;
+//       script.src = process.env.NEXT_PUBLIC_EMBED_SCRIPT_SRC;
+//       document.body.appendChild(script);
 
-      return () => {
-        document.body.removeChild(document.getElementById(process.env.NEXT_PUBLIC_EMBED_SCRIPT_ID));
-      };
-    }
-  }, [embedToken]);
+//       return () => {
+//         document.body.removeChild(document.getElementById(process.env.NEXT_PUBLIC_EMBED_SCRIPT_ID));
+//       };
+//     }
+//   }, [embedToken]);
 
-  return null;
-};
+//   return null;
+// };
 
-export default EmbedScriptLoader;
+// export default EmbedScriptLoader;
