+++ components/sliders/bridgeSlider.js
@@ -1,4 +1,4 @@
-import { useCustomSelector } from '@/customSelector/customSelector';
+import { useCustomSelector } from '@/customHooks/customSelector';
 import { getIconOfService, toggleSidebar } from '@/utils/utility';
 import { X } from 'lucide-react';
 import { usePathname, useRouter } from 'next/navigation';
